separated the boolean analyzer logic from CrosstabReducerHelper. Deleted the ReferenceDataMatcherAnalyzer as is identical with BooleanAnalyzerReducer
