--- components/reference-data/src/main/java/org/datacleaner/beans/ReferenceDataMatcherAnalyzerReducer.java
@@ -1,38 +1,0 @@
-/**
- * DataCleaner (community edition)
- * Copyright (C) 2014 Neopost - Customer Information Management
- *
- * This copyrighted material is made available to anyone wishing to use, modify,
- * copy, or redistribute it subject to the terms and conditions of the GNU
- * Lesser General Public License, as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
- * or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public License
- * for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this distribution; if not, write to:
- * Free Software Foundation, Inc.
- * 51 Franklin Street, Fifth Floor
- * Boston, MA  02110-1301  USA
- */
-package org.datacleaner.beans;
-
-import java.util.Collection;
-
-import org.datacleaner.api.AnalyzerResultReducer;
-
-public class ReferenceDataMatcherAnalyzerReducer implements AnalyzerResultReducer<BooleanAnalyzerResult> {
-
-    @Override
-    public BooleanAnalyzerResult reduce(Collection<? extends BooleanAnalyzerResult> partialResults) {
-        if (partialResults.isEmpty()) {
-            return null;
-        }
-        // the implementation of this reducer would be almost identical with
-        // BooleanAnalyzerReducer. 
-        final BooleanAnalyzerReducer booleanAnalyzerReducer = new BooleanAnalyzerReducer();
-        return booleanAnalyzerReducer.reduce(partialResults);
-    }
-}
